additional i18n support
Using `Flask-Babel`, internationalize most content in the app. There are
probably a few I missed, but this update should put most of the text in
`messages.pot`.

This will help us to add French text to all parts of the app, not just
the translation. It will also help in the future if/when we want to add
more languages.