support multiple languages
Allow all routes to be prepended with a `/<lang>/`. For example,
allow the `/about` page to be reached with `/about` and `/en/about`.
Currently the only options for `lang` are `en` and `fr`.

This is just a first pass. It doesn't actually add translations and
doesn't update any of the links in the application.