include locale in urls within application
Using `@app.url_defaults`, add the locale in links generated with
`url_for`. This means that once a user is within a locale namespace, all
links in the application will continue using that namespace. This change
also excludes the default locale, so when a user visits `/`, they will
be linked to `/about`, not `/en/about`.